Edit the Error component to be a dialog not a full page
